+++ src/main/java/flab/nutridiary/diary/domain/calculator/NutritionCalculator.java
@@ -1,6 +1,8 @@
-package flab.nutridiary.diary.domain;
+package flab.nutridiary.diary.domain.calculator;
 
 import flab.nutridiary.commom.exception.BusinessException;
+import flab.nutridiary.diary.domain.CalculatedNutrition;
+import flab.nutridiary.diary.domain.ProductIntakeInfo;
 import flab.nutridiary.product.domain.NutritionFacts;
 import flab.nutridiary.product.domain.Product;
 import flab.nutridiary.product.repository.ProductRepository;
